🚚(front) move InstructorView component
InstructorView component is only used by the InstructorWrapper component, so it
doesn't need to be in the root folder.
